+++ http.rest
@@ -18,5 +18,5 @@ Content-Type: application/json
 
 {
     "email": "Jill@Crown.lt",
-    "password": ""
+    "password": "secret123"
 }
